+++ driver-core/src/main/java/com/datastax/driver/core/Connection.java
@@ -309,7 +309,7 @@ class Connection extends org.apache.cassandra.transport.Connection
                     } else {
                         ce = new TransportException(address, "Error writing", writeFuture.getCause());
                     }
-                    handler.callback.onException(Connection.this, defunct(ce));
+                    handler.callback.onException(Connection.this, defunct(ce), System.nanoTime() - handler.startTime);
                 } else {
                     logger.trace("[{}] request sent successfully", name);
                 }
@@ -529,7 +529,7 @@ class Connection extends org.apache.cassandra.transport.Connection
                     return;
                 }
                 handler.cancelTimeout();
-                handler.callback.onSet(Connection.this, response);
+                handler.callback.onSet(Connection.this, response, System.nanoTime() - handler.startTime);
             }
         }
 
@@ -549,7 +549,8 @@ class Connection extends org.apache.cassandra.transport.Connection
             Iterator<ResponseHandler> iter = pending.values().iterator();
             while (iter.hasNext())
             {
-                iter.next().callback.onException(Connection.this, ce);
+                ResponseHandler handler = iter.next();
+                handler.callback.onException(Connection.this, ce, System.nanoTime() - handler.startTime);
                 iter.remove();
             }
         }
@@ -585,18 +586,18 @@ class Connection extends org.apache.cassandra.transport.Connection
         }
 
         @Override
-        public void onSet(Connection connection, Message.Response response, ExecutionInfo info) {
-            onSet(connection, response);
+        public void onSet(Connection connection, Message.Response response, ExecutionInfo info, long latency) {
+            onSet(connection, response, latency);
         }
 
         @Override
-        public void onSet(Connection connection, Message.Response response) {
+        public void onSet(Connection connection, Message.Response response, long latency) {
             this.address = connection.address;
             super.set(response);
         }
 
         @Override
-        public void onException(Connection connection, Exception exception) {
+        public void onException(Connection connection, Exception exception, long latency) {
             // If all nodes are down, we will get a null connection here. This is fine, if we have
             // an exception, consumers shouldn't assume the address is not null.
             if (connection != null)
@@ -605,7 +606,7 @@ class Connection extends org.apache.cassandra.transport.Connection
         }
 
         @Override
-        public void onTimeout(Connection connection) {
+        public void onTimeout(Connection connection, long latency) {
             assert connection != null; // We always timeout on a specific connection, so this shouldn't be null
             this.address = connection.address;
             super.setException(new ConnectionException(connection.address, "Operation Timeouted"));
@@ -618,9 +619,9 @@ class Connection extends org.apache.cassandra.transport.Connection
 
     interface ResponseCallback {
         public Message.Request request();
-        public void onSet(Connection connection, Message.Response response);
-        public void onException(Connection connection, Exception exception);
-        public void onTimeout(Connection connection);
+        public void onSet(Connection connection, Message.Response response, long latency);
+        public void onException(Connection connection, Exception exception, long latency);
+        public void onTimeout(Connection connection, long latency);
     }
 
     static class ResponseHandler {
@@ -628,7 +629,9 @@ class Connection extends org.apache.cassandra.transport.Connection
         public final Connection connection;
         public final int streamId;
         public final ResponseCallback callback;
+
         private final Timeout timeout;
+        private final long startTime;
 
         public ResponseHandler(Connection connection, ResponseCallback callback) throws BusyConnectionException {
             this.connection = connection;
@@ -637,6 +640,8 @@ class Connection extends org.apache.cassandra.transport.Connection
 
             long timeoutMs = connection.factory.getReadTimeoutMillis();
             this.timeout = timeoutMs <= 0 ? null : connection.factory.timer.newTimeout(onTimeoutTask(), timeoutMs, TimeUnit.MILLISECONDS);
+
+            this.startTime = System.nanoTime();
         }
 
         void cancelTimeout() {
@@ -652,7 +657,7 @@ class Connection extends org.apache.cassandra.transport.Connection
             return new TimerTask() {
                 @Override
                 public void run(Timeout timeout) {
-                    callback.onTimeout(connection);
+                    callback.onTimeout(connection, System.nanoTime() - startTime);
                     cancelHandler();
                 }
             };