+++ driver-core/src/main/java/com/datastax/driver/core/ResultSetFuture.java
@@ -70,7 +70,7 @@ public class ResultSetFuture extends SimpleFuture<ResultSet> {
         }
 
         @Override
-        public void onSet(Connection connection, Message.Response response, ExecutionInfo info) {
+        public void onSet(Connection connection, Message.Response response, ExecutionInfo info, long latency) {
             try {
                 switch (response.type) {
                     case RESULT:
@@ -137,19 +137,19 @@ public class ResultSetFuture extends SimpleFuture<ResultSet> {
         }
 
         @Override
-        public void onSet(Connection connection, Message.Response response) {
+        public void onSet(Connection connection, Message.Response response, long latency) {
             // This is only called for internal calls (i.e, when the callback is not wrapped in ResponseHandler),
             // so don't bother with ExecutionInfo.
-            onSet(connection, response, null);
+            onSet(connection, response, null, latency);
         }
 
         @Override
-        public void onException(Connection connection, Exception exception) {
+        public void onException(Connection connection, Exception exception, long latency) {
             setException(exception);
         }
 
         @Override
-        public void onTimeout(Connection connection) {
+        public void onTimeout(Connection connection, long latency) {
             // This is only called for internal calls (i.e, when the callback is not wrapped in ResponseHandler).
             // So just set an exception for the final result, which should be handled correctly by said internal call.
             setException(new ConnectionException(connection.address, "Operation Timeouted"));